Update day09.rs
Strategy for length heuristic combined with AABB collision detection.

This heuristic specializes to the AoC input generation and doesn't generalize to any input shape generation.

One of the points is always on a massive horizontal distance compared to adjacent points. If the input were rotated 90 degrees, this solution would stop working.

this means we don't need to compare every vertex to every other vertex, we can compare a handful to the others.

This idea can probably be taken further yet.